requirements/base.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-cors-headers==4.4.0
django-filter==24.3
django-storages==1.14.4
psycopg2-binary==2.9.10
celery==5.4.0
django-celery-beat==2.7.0
redis==5.2.1
Pillow==11.1.0
openpyxl==3.1.5
python-decouple==3.8
gunicorn==23.0.0
bleach==6.1.0
